debut de fonctionnement issue sidecomment
