+++ scripts/pushIssues.js
@@ -1,6 +1,6 @@
-
-
-
+/*
+ * Post a simple comment from the box on the bottom of the page
+ */
 $(function(){
   $('#sendIssue').on('click', function(e){
     e.preventDefault();
@@ -20,36 +20,36 @@ $(function(){
 
 
 /* send pullrequest*/
-function postPull(user, pass, title, body, head){
-  var uploadURL ="https://api.github.com/repos/whispyy/API-git/pulls";
-  console.log(uploadURL);
-  var data = 
-  JSON.stringify(
-    {
-      title: title, 
-      body: body,
-      head: head,
-      base: "master"
-    }
-  );
-  postJSON(uploadURL, data, user, pass);
-}
+// function postPull(user, pass, title, body, head){
+//   var uploadURL ="https://api.github.com/repos/whispyy/API-git/pulls";
+//   console.log(uploadURL);
+//   var data = 
+//   JSON.stringify(
+//     {
+//       title: title, 
+//       body: body,
+//       head: head,
+//       base: "master"
+//     }
+//   );
+//   postJSON(uploadURL, data, user, pass);
+// }
 
-function postPullFromIssue(user,pass, issue, head){
-  var uploadURL ="https://api.github.com/repos/whispyy/API-git/pulls";
-  var data = 
-  JSON.stringify(
-    {
-      issue : issue,
-      head : head,
-      base : "master"
-    }
-  );
-  Request.postJSON(uploadURL, data, user, pass);
-}
+// function postPullFromIssue(user,pass, issue, head){
+//   var uploadURL ="https://api.github.com/repos/whispyy/API-git/pulls";
+//   var data = 
+//   JSON.stringify(
+//     {
+//       issue : issue,
+//       head : head,
+//       base : "master"
+//     }
+//   );
+//   Request.postJSON(uploadURL, data, user, pass);
+// }
 
-$(function(){
-  sideComments.on('commentPosted', function( comment ){
-    postPull(user,pass,"test",comment,"commit: test");
-  });
-});
+// $(function(){
+//   sideComments.on('commentPosted', function( comment ){
+//     postPull(user,pass,"test",comment,"commit: test");
+//   });
+// });
